+++ src/api/routes/reports/timeseries/model.js
@@ -1,12 +1,12 @@
 /**
- * CogniCity Server /floods/timeseries data model
- * @module src/api/floods/timeseries/model
+ * CogniCity Server /reports/timeseries data model
+ * @module src/api/reports/timeseries/model
  **/
  import Promise from 'bluebird';
 
  /**
- * Methods to interact with flood layers in database
-  * @alias module:src/api/floods/timeseries/model
+ * Methods to interact with report layer in database
+  * @alias module:src/api/reports/timeseries/model
   * @param {Object} config Server configuration
   * @param {Object} db PG Promise database instance
   * @param {Object} logger Configured Winston logger instance
@@ -17,12 +17,11 @@ export default (config, db, logger) => ({
   // Get all flood reports for a given city
   count: (start, end) => new Promise((resolve, reject) => {
     // Setup query
-    let query = `SELECT ts, count(local_area) FROM
-    (SELECT (cognicity.rem_get_flood(ts)).local_area, ts
-      FROM  generate_series($1::timestamp with time zone,
-            $2::timestamp with time zone,'1 hour')
-      as series(ts)) output
-    GROUP BY ts ORDER BY ts`;
+    let query = `SELECT ts, count(r.pkey)
+    FROM generate_series($1::timestamp with time zone,
+    $2::timestamp with time zone, '1 hour') ts
+    LEFT JOIN cognicity.all_reports r
+    ON date_trunc('hour', r.created_at) = ts GROUP BY ts ORDER BY ts`;
 
     // Setup values
     let values = [start, end];